GreenMall 前端商城初始化

+++ greenmall-entity/src/main/java/com/yzf/greenmall/entity/Goods.java
@@ -1,6 +1,7 @@
 package com.yzf.greenmall.entity;
 
 import lombok.Data;
+import sun.rmi.runtime.Log;
 
 import javax.persistence.*;
 import java.io.Serializable;
@@ -28,7 +29,7 @@ public class Goods implements Serializable {
     private Date createTime;// 创建时间
     private Date lastUpdateTime;// 最后修改时间 last_update_time lastUpdateTime
     private Boolean valid;// 是否有效，逻辑删除用，true:有效 false:无效
-    private Integer salesVolume;// 销量
+    private Long salesVolume;// 销量
     @Transient
     private String category;
     @Transient
